fix: move types into runtime
