+++ src/runtime/composables/authjs/useAuth.ts
@@ -8,7 +8,7 @@ import { determineCallbackUrl } from '../../utils/url'
 import { makeCWN, joinPathToApiURLWN, navigateToAuthPageWN, getRequestURLWN } from '../../utils/callWithNuxt'
 import { _fetch } from '../../utils/fetch'
 import { isNonEmptyObject } from '../../utils/checkSessionResult'
-import { CommonUseAuthReturn, SignOutFunc, GetSessionFunc, SignInFunc } from '../../../types'
+import { CommonUseAuthReturn, SignOutFunc, GetSessionFunc, SignInFunc } from '../../types'
 import { useTypedBackendConfig } from '../../helpers'
 import type { SessionData } from './useAuthState'
 import { createError, useNuxtApp, useRuntimeConfig, useRequestHeaders, useAuthState } from '#imports'
