+++ src/runtime/composables/authjs/useAuthState.ts
@@ -1,5 +1,5 @@
 import type { Session } from 'next-auth'
-import { CommonUseAuthStateReturn } from '../../../types'
+import { CommonUseAuthStateReturn } from '../../types'
 import { makeCommonAuthState } from '../commonAuthState'
 
 export type SessionData = Session
