+++ src/runtime/composables/commonAuthState.ts
@@ -2,7 +2,7 @@ import { computed } from 'vue'
 import getURL from 'requrl'
 import { joinURL } from 'ufo'
 import { useRuntimeConfig, useRequestEvent } from '#app'
-import { SessionLastRefreshedAt, SessionStatus } from '../../types'
+import { SessionLastRefreshedAt, SessionStatus } from '../types'
 import { useState } from '#imports'
 
 export const makeCommonAuthState = <SessionData>() => {
