+++ src/runtime/composables/local/useAuthState.ts
@@ -1,6 +1,6 @@
 import { computed, watch, ComputedRef } from 'vue'
 import { CookieRef } from '#app'
-import { CommonUseAuthStateReturn } from '../../../types'
+import { CommonUseAuthStateReturn } from '../../types'
 import { makeCommonAuthState } from '../commonAuthState'
 import { useTypedBackendConfig } from '../../helpers'
 import { useRuntimeConfig, useCookie, useState } from '#imports'
